Task1/requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
scikit-learn
scipy
opencv-python
tqdm
